refactor : 면접 도메인 테이블 컬럼 수정 및 api요청 수정 / feat : 면접 담당자 도메인 Command/Query 기능 구현

+++ src/main/java/com/piveguyz/empickbackend/employment/interviews/interviewSheetItem/query/dto/InterviewSheetItemQueryDTO.java
@@ -15,7 +15,7 @@ public class InterviewSheetItemQueryDTO {
     private Integer id;
     private Integer sheetId;
     private Integer criteriaId;
-    private Integer weight;
+    private Double weight;
     private Integer memberId;
     private LocalDateTime updatedAt;
 }
